refactor: move setup_logging to the package
