+++ scripts/dataset_download.py
@@ -1,4 +1,5 @@
 import argparse
+import logging
 import shutil
 from pathlib import Path
 
@@ -6,6 +7,9 @@ import yaml
 from huggingface_hub import snapshot_download
 
 from oxspires_tools.dataset import check_image_lidar_sync
+from oxspires_tools.utils import setup_logging
+
+logger = logging.getLogger(__name__)
 
 
 def load_config(config_path: str) -> dict:
@@ -22,13 +26,13 @@ def download_patterns(
     unpack: bool = False,
 ) -> list:
     """Download patterns from the HuggingFace repository."""
-    print(f"Repository: {repo_id}")
-    print(f"Local directory: {local_dir}")
-    print(f"Unpack archives: {unpack}")
-    print(f"Downloading {len(patterns)} pattern(s)...\n")
+    logger.info(f"Repository: {repo_id}")
+    logger.info(f"Local directory: {local_dir}")
+    logger.info(f"Unpack archives: {unpack}")
+    logger.info(f"Downloading {len(patterns)} pattern(s)...\n")
 
     for i, pattern in enumerate(patterns, 1):
-        print(f"[{i}/{len(patterns)}] {pattern}")
+        logger.info(f"[{i}/{len(patterns)}] {pattern}")
         snapshot_download(
             repo_id=repo_id,
             allow_patterns=pattern,
@@ -36,14 +40,14 @@ def download_patterns(
             repo_type=repo_type,
             use_auth_token=False,
         )
-        print(f"✅ Downloaded: {pattern}\n")
-    print("🏁 All downloads complete!")
+        logger.info(f"✅ Downloaded: {pattern}\n")
+    logger.info("🏁 All downloads complete!")
 
 
 def unpack_zip_files(local_dir: str):
     zip_files = list(Path(local_dir).rglob("*.zip"))
     for zip_file in zip_files:
-        print(f"Unzipping {zip_file}")
+        logger.info(f"Unzipping {zip_file}")
         shutil.unpack_archive(zip_file, extract_dir=zip_file.parent)
         zip_file.unlink()
 
@@ -61,6 +65,7 @@ def get_args():
 
 
 def main():
+    setup_logging()
     args = get_args()
 
     config = load_config(args.config)
@@ -80,7 +85,7 @@ def main():
         for seq_dir in sequences.iterdir():
             if not seq_dir.is_dir():
                 continue
-            print(f"\nChecking sequence: {seq_dir.name}")
+            logger.info(f"\nChecking sequence: {seq_dir.name}")
             check_image_lidar_sync(
                 image_dir=Path(seq_dir) / "raw" / "cam0",
                 lidar_dir=Path(seq_dir) / "processed" / "vilens-slam" / "undist-clouds",
